oeqa/selftest/tinfoil: Fix intermittent event loss issue in test
We've been seeing occasional test failures on the autobuilder where
we don't see the expected events. It turns out this is due to
run_command being helpful and eating them if the server is fast and
the client slow. Adding a sleep into the run_command code makes the
failure consistent.

Use a new "handle_events" argument to allow us to handle all the
events which is what this test requires.

[YOCTO #14585]

(From OE-Core rev: b97f64992a1791abe56d9324ece1950cfa9fd2d2)

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>